refactor: extract chat-template rendering into standalone dynamo-prompt crate
Moves chat-template / prompt formatting out of lib/llm into a new standalone,
runtime-free `dynamo-prompt` crate so it can be reused by external OpenAI
frontends (e.g. an SGLang Rust HTTP server) without the Dynamo runtime.

The renderer is a *bridge* between OpenAI request types (`dynamo-protocols`)
and the HF `chat_template` jinja engine (`minijinja`); it shares no code with
the tokenizer itself. So rather than fold it into `dynamo-tokenizers` (which
would force a tokenizers->protocols dependency and orphan-rule/dev-cycle
gymnastics for zero code-sharing benefit), it gets its own crate that owns the
`OAIChatLikeRequest` trait and depends on both leaves.

- New `dynamo-prompt` crate: HF chat_template rendering (minijinja+pycompat),
  DeepSeek native formatters, tokenizer_config parsing, and the
  `OAIChatLikeRequest` trait. The default impl for `dynamo-protocols`' OpenAI
  chat request is orphan-rule-legal here (the trait is local). Re-exports
  `dynamo-tokenizers` as a one-import convenience.
- `dynamo-tokenizers` returns to a pure tokenization leaf: no protocols, no
  minijinja, no feature flags.
- `dynamo-protocols` stays a pure wire-types leaf: no tokenizers, no minijinja.
- lib/llm depends on `dynamo-prompt`; keeps its `Nv*` impls, the
  `prompt_formatter_from_mdc` adapter, and media-IO via the local
  `MediaRequestExt` trait. `PromptContextMixin` moves to the crate;
  model_card re-exports it.

Each crate now has one job and dependencies flow one way
(dynamo-prompt -> {protocols, tokenizers}; lib/llm -> all three). No
dependency cycle, no feature gymnastics.

Verified: dynamo-prompt 87 tests; dynamo-tokenizers 25 tests (lean); dynamo-llm
builds + preprocessor 41 + deepseek 16 tests pass; clippy/fmt clean. Bindings
rebuilt via maturin. E2E on 2x L40S: dynamo.frontend + dynamo.sglang Qwen3-0.6B
renders the Qwen chat template correctly (multi-turn /no_think -> finish stop;
tool-calling turn renders the tools section and emits a valid <tool_call>).
