+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "lib/llm",
     "lib/runtime",
     "lib/tokenizers",
+    "lib/prompt",
     "lib/tokens",
     "lib/data-gen",
     "lib/kv-hashing",
@@ -44,6 +45,7 @@ keywords = ["llm", "genai", "inference", "nvidia", "distributed"]
 dynamo-runtime = { path = "lib/runtime", version = "1.2.0" }
 dynamo-llm = { path = "lib/llm", version = "1.2.0" }
 dynamo-tokenizers = { path = "lib/tokenizers", version = "1.2.0" }
+dynamo-prompt = { path = "lib/prompt", version = "1.2.0" }
 dynamo-tokens = { path = "lib/tokens", version = "1.2.0" }
 dynamo-data-gen = { path = "lib/data-gen", version = "1.2.0" }
 dynamo-kv-hashing = { path = "lib/kv-hashing", version = "1.2.0" }
