+++ lib/prompt/src/deepseek_common.rs
